added updateUser mutation
