+++ api/src/__tests__/db.test.js
@@ -16,6 +16,7 @@ describe('User', () => {
         resetTokenExpiresAt: null,
         newsGeneral: true,
         newsBusiness: false,
+        newsEntertainment: false,
         newsHealth: false,
         newsScience: false,
         newsSports: false,
@@ -39,6 +40,7 @@ describe('User', () => {
       resetTokenExpiresAt: null,
       newsGeneral: true,
       newsBusiness: false,
+      newsEntertainment: false,
       newsHealth: true,
       newsScience: false,
       newsSports: false,
@@ -76,6 +78,7 @@ describe('User', () => {
           resetTokenExpiresAt: null,
           newsGeneral: true,
           newsBusiness: false,
+          newsEntertainment: false,
           newsHealth: false,
           newsScience: false,
           newsSports: false,
@@ -100,6 +103,7 @@ describe('User', () => {
           resetTokenExpiresAt: null,
           newsGeneral: true,
           newsBusiness: false,
+          newsEntertainment: false,
           newsHealth: true,
           newsScience: false,
           newsSports: false,
@@ -130,6 +134,7 @@ describe('User', () => {
           resetTokenExpiresAt: null,
           newsGeneral: true,
           newsBusiness: false,
+          newsEntertainment: false,
           newsHealth: true,
           newsScience: false,
           newsSports: false,
